feat: implement password generation functionality and update command flags
